requirements.txt
langchain
deeplake
openai
tiktoken
python-dotenv
